Enum novatel_oem6_api::ResponseID

pub enum ResponseID {
    Ok,
    LogInvalid,
    OutOfResources,
    PacketNotVerified,
    CommandFailed,
    InvalidID,
    InvalidField,
    InvalidChecksum,
    MissingField,
    ArrayOverflow,
    ErrorField,
    InvalidTrigger,
    AuthcodeOverflow,
    InvalidDate,
    InvalidAuthcode,
    NoModel,
    InvalidModel,
    InvalidChannel,
    InvalidRate,
    NoMask,
    LockedChannels,
    InvalidTime,
    InvalidPort,
    InvalidMessage,
    InvalidPRN,
    PRNNotLocked,
    PRNLockoutOverflow,
    PRNAlreadyLocked,
    Timeout,
    UnknownPort,
    BadHex,
    InvalidBaud,
    ModelInvalidMessage,
    RequiresFailMode,
    InvalidOffset,
    MessageOverflow,
    PreciseTimeAlreadyKnown,
    Unknown,
}

Response values returned after sending a command to the device.

Error values will be returned as part of the OEMError::CommandError variant.

Variants

Ok

Command was received correctly

LogInvalid

Requested log type does not exist

OutOfResources

The request has exceeded some (unspecified) limit

PacketNotVerified

Data packet is not verified

CommandFailed

Command was attempted, but failed

InvalidID

Input message ID is not valid

InvalidField

An input field was invalid. See the OEMError::CommandError variant's description field for information about the specific field that caused the error.

InvalidChecksum

The checksum of the sent message was invalid

MissingField

A field is missing from the sent message

ArrayOverflow

An input field contains more array elements than allowed. See the OEMError::CommandError variant's description field for information about the specific field that caused the error.

ErrorField

An input field's value is outside acceptable limits. See the OEMError::CommandError variant's description field for information about the specific field that caused the error.

InvalidTrigger

The requested trigger type is invalid for the requested log type

AuthcodeOverflow

Too many authcodes are stored in the receiver. The receiver firmware must be reloaded

InvalidDate

This error is related to the inputting of authcodes. Indicates the date attached to the code is not valid

InvalidAuthcode

The authcode entered is not valid

NoModel

The model requested for removal does not exist

InvalidModel

The model attached to the authcode is not valid

InvalidChannel

The selected channel is invalid

InvalidRate

The requested rate is invalid

NoMask

The word has no mask for this type of log

LockedChannels

Channels are locked due to an error

InvalidTime

The injected time is invalid

InvalidPort

The COM/USB port is not supported

InvalidMessage

The sent message is invalid

InvalidPRN

The PRN is invalid

PRNNotLocked

The PRN is not locked out

PRNLockoutOverflow

The PRN lockout list is full

PRNAlreadyLocked

The PRN is already locked out

Timeout

Message timed out

UnknownPort

Unknown COM/USB port requested

BadHex

Hex string not formatted correctly

InvalidBaud

The baud rate is invalid

ModelInvalidMessage

The sent message is invalid for this model of receiver

RequiresFailMode

Command is only valid if NVM is in fail mode

InvalidOffset

The offset is invalid

MessageOverflow

The maximum number of user messages has been reached

PreciseTimeAlreadyKnown

GPS precise time is already known

Unknown

Catch-all value for unknown response IDs